requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
PyQt5>=5.15.0
scikit-image>=0.21.0
